Move the lexer to the common directory
We want to be able to re-use it for the assemblers
