+++ common/ROMakefile
@@ -2,7 +2,7 @@
 
 COMPONENT = common
 
-OBJS = error utils buffer ir bitset builtins constant_pool string_pool type sizet_vector vm_heap
+OBJS = keywords lexer error stream utils buffer ir bitset builtins constant_pool string_pool type sizet_vector vm_heap
 
 CFLAGS ?= -Wxla
