+++ common/lexer.c
