+++ common/stream.c
