+++ common/stream.h
